+++ CHANGELOG.md
@@ -5,3 +5,4 @@
 ### Features Added
 - Upgrade Calico to 3.12.0 (PR [#2253](https://github.com/scality/metalk8s/pull/2253))
 - Upgrade `containerd` to 1.2.13 (PR [#2369](https://github.com/scality/metalk8s/pull/2369))
+- Enable `seccomp` support in `containerd` (Issue [#2259](https://github.com/scality/metalk8s/issues/2259), PR [#2369](https://github.com/scality/metalk8s/pull/2369))